requirements.txt
autopep8==1.5.4
coverage==5.2.1
Django==3.0.8
djangorestframework==3.11.0
django-extensions==2.2.8
django-rest-swagger==2.2.0
flake8==3.8.3
isort==5.9.3
model-mommy==2.0.0
psycopg2-binary==2.8.4
pycodestyle==2.6.0